Fix: board 삭제/수정 버튼 토글 기능으로 변경
